Time bugfix

+++ storage/cachestatistics.hpp
@@ -36,7 +36,7 @@ struct CacheStatistics {
     uint32_t m_flushed;
     uint32_t m_lookups;
     uint32_t m_lookups2;
-    uint32_t m_put_time;
+    uint64_t m_put_time;
     CacheStatistics operator-(const CacheStatistics& o) const noexcept;
     friend std::ostream& operator<<(std::ostream& os, const CacheStatistics& statistics) noexcept;
 };
